requirements.txt
# ===== CORE DJANGO DEPENDENCIES =====
Django==4.2.16
django-htmx==1.17.2
django-csp==3.8
django-ratelimit==4.1.0
python-dotenv==1.0.0

# ===== DATABASE =====
psycopg2-binary==2.9.9

# ===== API & HTTP =====
requests==2.31.0

# ===== IMAGE PROCESSING (REQUIRED for ImageField) =====
# Pillow is REQUIRED for BlogPost.featured_image and Tool.image
Pillow==10.4.0

# ===== NEWSLETTER DEPENDENCIES =====
# Para procesamiento de emails y newsletters
django-allauth==0.61.1  # Si usas autenticación social
bleach==6.1.0          # Para limpiar HTML en emails

# ===== PRODUCTION DEPENDENCIES =====
# Uncomment when deploying to production
# gunicorn==21.2.0
# whitenoise==6.6.0
# django-storages==1.14.2  # Para S3/Cloud storage
# boto3==1.35.0           # Para AWS S3

# ===== DEVELOPMENT DEPENDENCIES =====
# Uncomment for development
# django-debug-toolbar==4.2.0
# django-extensions==3.2.3

coverage==7.9.1

psycopg2-binary>=2.9.0
